Fix top_k type

+++ config.py
@@ -176,7 +176,7 @@ class GPTSoVitsConfig(AsDictMixin):
     lang: str = "auto"
     format: str = "wav"
     segment_size: int = 50
-    top_k: float = 5
+    top_k: int = 5
     top_p: float = 1
     temperature: float = 1
     presets: Dict[str, GPTSoVitsPreset] = field(default_factory=lambda: {"default": GPTSoVitsPreset(),
